+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.23] 2021-12-11
 
+### Changed
 
+- The common implementation of the memory manager as a subsystem was excluded. Now local usage of allocators and
+memory arenas is more preffered way.
 
 ## [0.5.22] 2021-12-01